Combine test_utils_.py with test_amo_utils.py (bug 1037132)
