+++ apps/amo/tests/test_amo_utils.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
+import collections
 import os
 import tempfile
 import unittest
+
 
 from django.conf import settings
 from django.core.cache import cache
@@ -9,13 +11,16 @@ from django.core.validators import ValidationError
 from django.utils import translation
 
 import mock
-from nose.tools import eq_, assert_raises, raises
+from nose.tools import assert_raises, eq_, ok_, raises
 from product_details import product_details
 
-from amo.utils import (cache_ns_key, escape_all, find_language,
-                       LocalFileStorage, no_translation, resize_image,
-                       rm_local_tmp_dir, slugify, slug_validator, to_language)
-from mkt.translations.models import Translation
+import amo
+from amo.utils import (attach_trans_dict, cache_ns_key, escape_all,
+                       find_language, LocalFileStorage, no_translation,
+                       resize_image, rm_local_tmp_dir, slugify, slug_validator,
+                       to_language)
+from mkt.webapps.models import Addon
+
 
 u = u'Ελληνικά'
 
@@ -35,18 +40,19 @@ def test_slugify():
     def check(x, y):
         eq_(slugify(x), y)
         slug_validator(slugify(x))
-    s = [('xx x  - "#$@ x', 'xx-x-x'),
-         (u'Bän...g (bang)', u'bäng-bang'),
-         (u, u.lower()),
-         (x, x.lower()),
-         (y, x.lower()),
-         ('    a ', 'a'),
-         ('tags/', 'tags'),
-         ('holy_wars', 'holy_wars'),
-         # I don't really care what slugify returns.  Just don't crash.
-         (u'x荿', u'x\u837f'),
-         (u'ϧ΃蒬蓣', u'\u03e7\u84ac\u84e3'),
-         (u'¿x', u'x'),
+    s = [
+        ('xx x  - "#$@ x', 'xx-x-x'),
+        (u'Bän...g (bang)', u'bäng-bang'),
+        (u, u.lower()),
+        (x, x.lower()),
+        (y, x.lower()),
+        ('    a ', 'a'),
+        ('tags/', 'tags'),
+        ('holy_wars', 'holy_wars'),
+        # I don't really care what slugify returns. Just don't crash.
+        (u'x荿', u'x\u837f'),
+        (u'ϧ΃蒬蓣', u'\u03e7\u84ac\u84e3'),
+        (u'¿x', u'x'),
     ]
     for val, expected in s:
         yield check, val, expected
@@ -159,7 +165,7 @@ class TestLocalFileStorage(unittest.TestCase):
         dp = os.path.join(self.tmp, 'path', 'to')
         self.stor.open(os.path.join(dp, 'file.txt'), 'w').close()
         assert os.path.exists(self.stor.path(dp)), (
-                                        'Directory not created: %r' % dp)
+            'Directory not created: %r' % dp)
 
     def test_do_not_make_file_dirs_when_reading(self):
         fpath = os.path.join(self.tmp, 'file.txt')
@@ -265,3 +271,86 @@ class TestEscapeAll(unittest.TestCase):
         eq_(res['dict'], {'x': expected})
         eq_(res['list'], [expected])
         eq_(res['bool'], True)
+
+
+class TestAttachTransDict(amo.tests.TestCase):
+    """
+    Tests for attach_trans_dict. For convenience, we re-use Addon model instead
+    of mocking one from scratch and we rely on internal Translation unicode
+    implementation, because mocking django models and fields is just painful.
+    """
+
+    def test_basic(self):
+        addon = amo.tests.addon_factory(
+            name='Name', description='Description <script>alert(42)</script>!',
+            homepage='http://home.pa.ge', privacy_policy='Policy',
+            support_email='sup@example.com', support_url='http://su.pport.url')
+        addon.save()
+
+        # Quick sanity checks: is description properly escaped? The underlying
+        # implementation should leave localized_string un-escaped but never use
+        # it for __unicode__. We depend on this behaviour later in the test.
+        ok_('<script>' in addon.description.localized_string)
+        ok_(not '<script>' in addon.description.localized_string_clean)
+        ok_(not '<script>' in unicode(addon.description))
+
+        # Attach trans dict.
+        attach_trans_dict(Addon, [addon])
+        ok_(isinstance(addon.translations, collections.defaultdict))
+        translations = dict(addon.translations)
+
+        # addon.translations is a defaultdict.
+        eq_(addon.translations['whatever'], [])
+
+        # No-translated fields should be absent.
+        ok_(None not in translations)
+
+        # Build expected translations dict.
+        expected_translations = {
+            addon.privacy_policy_id: [
+                ('en-us', unicode(addon.privacy_policy))],
+            addon.description_id: [
+                ('en-us', unicode(addon.description))],
+            addon.homepage_id: [('en-us', unicode(addon.homepage))],
+            addon.name_id: [('en-us', unicode(addon.name))],
+            addon.support_email_id: [('en-us', unicode(addon.support_email))],
+            addon.support_url_id: [('en-us', unicode(addon.support_url))]
+        }
+        eq_(translations, expected_translations)
+
+    def test_multiple_objects_with_multiple_translations(self):
+        addon = amo.tests.addon_factory()
+        addon.description = {
+            'fr': 'French Description',
+            'en-us': 'English Description'
+        }
+        addon.save()
+        addon2 = amo.tests.addon_factory(description='English 2 Description')
+        addon2.name = {
+            'fr': 'French 2 Name',
+            'en-us': 'English 2 Name',
+            'es': 'Spanish 2 Name'
+        }
+        addon2.save()
+        attach_trans_dict(Addon, [addon, addon2])
+        eq_(set(addon.translations[addon.description_id]),
+            set([('en-us', 'English Description'),
+                 ('fr', 'French Description')]))
+        eq_(set(addon2.translations[addon2.name_id]),
+            set([('en-us', 'English 2 Name'),
+                 ('es', 'Spanish 2 Name'),
+                 ('fr', 'French 2 Name')]))
+
+
+def test_has_links():
+    html = 'a text <strong>without</strong> links'
+    assert not amo.utils.has_links(html)
+
+    html = 'a <a href="http://example.com">link</a> with markup'
+    assert amo.utils.has_links(html)
+
+    html = 'a http://example.com text link'
+    assert amo.utils.has_links(html)
+
+    html = 'a badly markuped <a href="http://example.com">link'
+    assert amo.utils.has_links(html)
